+++ backend/src/main/java/com/cs308_team_3/sabanci_pharmacy/entity/Address.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Data;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
+import java.math.BigInteger;
+
 @Entity
 @Table(name = "addresses")
 @Data
@@ -28,7 +30,7 @@ public class Address {
     private Boolean isDefault = false;
 
     @Column
-    private Integer phone;
+    private BigInteger phone;
 
     // Many addresses belong to one user
     @ManyToOne(fetch = FetchType.LAZY)
